requirements.txt
python-dotenv
requests
websockets
python-telegram-bot
pytz
